chore: add authorization admin access to request update delete accept reject and return

+++ src/main/java/com/bootcamp/bookrentalsystem/controller/RequestController.java
@@ -1,8 +1,10 @@
 package com.bootcamp.bookrentalsystem.controller;
 
+import com.bootcamp.bookrentalsystem.exception.ForbiddenException;
 import com.bootcamp.bookrentalsystem.model.Request;
 import com.bootcamp.bookrentalsystem.model.User;
 import com.bootcamp.bookrentalsystem.service.BookService;
+import com.bootcamp.bookrentalsystem.service.JwtService;
 import com.bootcamp.bookrentalsystem.service.RequestService;
 import com.bootcamp.bookrentalsystem.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -19,12 +21,14 @@ public class RequestController {
     private final BookService bookService;
     private final UserService userService;
     private final RequestService requestService;
+    private final JwtService jwtService;
 
     @Autowired
-    public RequestController(BookService bookService, UserService userService, RequestService requestService) {
+    public RequestController(BookService bookService, UserService userService, RequestService requestService, JwtService jwtService) {
         this.bookService = bookService;
         this.userService = userService;
         this.requestService = requestService;
+        this.jwtService = jwtService;
     }
 
     @PostMapping()
@@ -42,13 +46,23 @@ public class RequestController {
     }
 
     @PatchMapping("/{requestId}")
-    public ResponseEntity<Request> updateRequest(@PathVariable Long requestId, @RequestBody Request updatedRequest) {
+    public ResponseEntity<Request> updateRequest(@RequestHeader("Authorization") String token, @PathVariable Long requestId, @RequestBody Request updatedRequest) {
+        // Validate and decode the JWT token
+        if (!jwtService.isValidAdminToken(token)) {
+            throw new ForbiddenException("Access Denied!!");
+        }
+
         Request request = this.requestService.updateRequestById(requestId, updatedRequest);
         return new ResponseEntity<>(request, HttpStatus.OK);
     }
 
     @DeleteMapping("/{requestId}")
-    public ResponseEntity<Map<String, Boolean>> deleteRequestById(@PathVariable Long requestId) {
+    public ResponseEntity<Map<String, Boolean>> deleteRequestById(@RequestHeader("Authorization") String token, @PathVariable Long requestId) {
+        // Validate and decode the JWT token
+        if (!jwtService.isValidAdminToken(token)) {
+            throw new ForbiddenException("Access Denied!!");
+        }
+
         Map<String, Boolean> response = requestService.deleteRequestById(requestId);
         return new ResponseEntity<>(response, HttpStatus.OK);
     }
@@ -59,19 +73,34 @@ public class RequestController {
     }
 
     @PatchMapping("/{requestId}/accept")
-    public ResponseEntity<Request> acceptRequest(@PathVariable Long requestId) {
+    public ResponseEntity<Request> acceptRequest(@RequestHeader("Authorization") String token, @PathVariable Long requestId) {
+        // Validate and decode the JWT token
+        if (!jwtService.isValidAdminToken(token)) {
+            throw new ForbiddenException("Access Denied!!");
+        }
+
         Request acceptedRequest = requestService.acceptRequest(requestId);
         return new ResponseEntity<>(acceptedRequest, HttpStatus.OK);
     }
 
     @PatchMapping("/{requestId}/reject")
-    public ResponseEntity<Request> rejectRequest(@PathVariable Long requestId) {
+    public ResponseEntity<Request> rejectRequest(@RequestHeader("Authorization") String token, @PathVariable Long requestId) {
+        // Validate and decode the JWT token
+        if (!jwtService.isValidAdminToken(token)) {
+            throw new ForbiddenException("Access Denied!!");
+        }
+
         Request rejectedRequest = requestService.rejectRequest(requestId);
         return new ResponseEntity<>(rejectedRequest, HttpStatus.OK);
     }
 
     @PatchMapping("/{requestId}/return")
-    public ResponseEntity<String> returnBook(@PathVariable Long requestId) {
+    public ResponseEntity<String> returnBook(@RequestHeader("Authorization") String token, @PathVariable Long requestId) {
+        // Validate and decode the JWT token
+        if (!jwtService.isValidAdminToken(token)) {
+            throw new ForbiddenException("Access Denied!!");
+        }
+
         Request archivedRequest = requestService.returnBook(requestId);
         return ResponseEntity.ok("Book returned successfully.");
     }
